Aligning aggregation with API
Have been reworking the API surface for blog posts over the past week. Now updating blog post aggregation. Decided to change API surface to return empty collections where database value is NULL. Also to expect empty collections as parameters, not NULL.

+++ routes/utility.js
@@ -160,7 +160,9 @@ router.get( '/images/:url', async ( req, res ) => {
   }
 
   // Clean up
-  fs.unlinkSync( local );
+  if( fs.existsSync( path ) ) {
+    fs.unlinkSync( local );
+  }
 
   // Return array of URL strings
   res.json( results );
